+++ modules/Users/Search.inc.php
@@ -90,8 +90,11 @@ if ( ! $_REQUEST['search_modfunc'])
 
 			if ( User('PROFILE') === 'admin' )
 			{
-				//FJ if only one school, no Search All Schools option
-				if ( SchoolInfo( 'SCHOOLS_NB' ) > 1 )
+				// FJ if only one school, no Search All Schools option.
+				// Restrict Search All Schools to user schools.
+				if ( SchoolInfo( 'SCHOOLS_NB' ) > 1
+					&& ( ! trim( User( 'SCHOOLS' ), ',' )
+						|| mb_substr_count( User( 'SCHOOLS' ), ',' ) > 2 ) )
 				{
 					echo '<label><input type="checkbox" name="_search_all_schools" value="Y"' .
 						( Preferences( 'DEFAULT_ALL_SCHOOLS' ) == 'Y' ? ' checked' : '' ) . '>&nbsp;' .