+++ CHANGES.md
@@ -6,6 +6,9 @@ Changes in 2.9.13
 - Always use SchoolInfo() instead of querying SCHOOLS DB table, programwide
 - Unset current student after setting new current school in Schools.php & CopySchool.php
 - Check if Update() version < ROSARIO_VERSION in Update.fnc.php
+- Admin Schools restriction in Profile.php, Exceptions.php, User.php & Geneal_Info.inc.php (sponsored by Aptiris)
+- Update database for Admin Schools restriction in rosariosis.sql & Update.fnc.php
+- Restrict Search All Schools to user schools (sponsored by Aptiris)
 
 Changes in 2.9.12
 -----------------