+++ CHANGES.md
@@ -1,12 +1,16 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 2.9.14
+-----------------
+- Add "No courses found" error in Side.php
+
 Changes in 2.9.13
 -----------------
 - Always use SchoolInfo() instead of querying SCHOOLS DB table, programwide
 - Unset current student after setting new current school in Schools.php & CopySchool.php
 - Check if Update() version < ROSARIO_VERSION in Update.fnc.php
-- Admin Schools restriction in Profile.php, Exceptions.php, User.php & Geneal_Info.inc.php (sponsored by Aptiris)
+- Admin Schools restriction in Profile.php, Exceptions.php, User.php & General_Info.inc.php (sponsored by Aptiris)
 - Update database for Admin Schools restriction in rosariosis.sql & Update.fnc.php
 - Restrict Search All Schools to user schools (sponsored by Aptiris)